Move unlock_submitted_document_form from meeting to locking module.
This is in prevision of generalising that form to also use it
to unlock documents copied to a workspace.
