+++ opengever/locking/browser/templates/unlock_submitted_document_form.pt
@@ -4,7 +4,7 @@
       xmlns:i18n="http://xml.zope.org/namespaces/i18n"
       lang="en"
       metal:use-macro="context/main_template/macros/master"
-      i18n:domain="opengever.meeting">
+      i18n:domain="opengever.locking">
 
 <body>
